requirements.txt
# Flask and API
Flask==3.0.0
flask-cors==4.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1

# Machine Learning
scikit-learn==1.3.2

# Google Gemini API
google-generativeai==0.4.1

# Social Media APIs
praw==7.7.1
google-api-python-client==2.108.0
atproto==0.0.38
pytrends==4.9.2

# Utilities
python-dateutil==2.8.2
requests==2.31.0
tqdm==4.66.1